CORE-455: review findings
