+++ backend/src/main/java/quarano/department/web/TrackedCaseController.java
@@ -13,7 +13,6 @@ import quarano.core.web.LoggedIn;
 import quarano.core.web.MappedPayloads;
 import quarano.department.CaseType;
 import quarano.department.EnrollmentCompletion;
-import quarano.department.RegistrationManagement;
 import quarano.department.TrackedCase;
 import quarano.department.TrackedCase.TrackedCaseIdentifier;
 import quarano.department.TrackedCaseProperties;
@@ -26,7 +25,6 @@ import quarano.department.web.TrackedCaseRepresentations.ValidatedContactCase;
 import quarano.department.web.TrackedCaseRepresentations.ValidatedIndexCase;
 import quarano.diary.DiaryManagement;
 import quarano.tracking.TrackedPerson;
-import quarano.tracking.TrackedPersonRepository;
 import quarano.tracking.web.TrackedPersonDto;
 import quarano.tracking.web.TrackingController;
 
@@ -71,9 +69,6 @@ public class TrackedCaseController {
 	private final @NonNull TrackedCaseProperties configuration;
 	private final @NonNull TrackedCaseRepresentations representations;
 	private final @NonNull HealthDepartments rkiDepartments;
-	private final @NonNull TrackedPersonRepository people;
-	private final @NonNull RegistrationManagement registrationManagement;
-	private final @NonNull RegistrationRepresentations registrationRepresentations;
 
 
 	@GetMapping(path = "/hd/cases")
@@ -229,24 +224,6 @@ public class TrackedCaseController {
 				.map(cases::save)
 				.map(it -> representations.toRepresentation(it))
 				.concludeIfValid(ResponseEntity::ok);
-	}
-
-	@DeleteMapping("/hd/cases/{identifier}/account")
-	HttpEntity<?> removeAccountOfTrackedPerson(@PathVariable TrackedCaseIdentifier identifier, @LoggedIn Account account) {
-
-		return cases.findById(identifier)
-				.filter(it -> it.belongsTo(account.getDepartmentId()))
-				.filter(it -> it.getTrackedPerson().getAccount().isPresent())
-				.map(it -> {
-					TrackedPerson trackedPerson = it.getTrackedPerson().deleteAccountRegistration();
-					people.save(trackedPerson);
-
-					return registrationManagement.initiateRegistration(it).toTry()
-							.map(at -> registrationRepresentations.toRepresentation(at, it))
-							.fold(at -> ResponseEntity.badRequest().body(at.getMessage()),
-									ResponseEntity::ok);
-				})
-				.orElseGet(() -> ResponseEntity.badRequest().build());
 	}
 
 	@GetMapping("/enrollments")
