+++ backend/src/main/java/quarano/department/web/TrackedCaseStatusAware.java
@@ -6,9 +6,6 @@ import static quarano.department.web.TrackedCaseLinkRelations.*;
 import lombok.Getter;
 import quarano.core.EnumMessageSourceResolvable;
 import quarano.department.TrackedCase;
-import quarano.department.TrackedCase.Status;
-
-import java.util.function.Supplier;
 
 import org.springframework.context.support.MessageSourceAccessor;
 import org.springframework.hateoas.IanaLinkRelations;
@@ -39,18 +36,15 @@ public class TrackedCaseStatusAware<T extends RepresentationModel<T>> extends Re
 
 		var caseId = trackedCase.getId();
 		var controller = on(TrackedCaseController.class);
-
-		Supplier<Object> uri = () -> on(RegistrationController.class).createRegistration(caseId, null);
+		var registrationController = on(RegistrationController.class);
 
 		return Links.of(MvcLink.of(controller.concludeCase(caseId, null), CONCLUDE))
 				.and(MvcLink.of(controller.getDiaryOfCase(caseId, null), DIARY))
 				.and(MvcLink.of(controller.getCase(caseId, null), IanaLinkRelations.SELF))
 				.andIf(trackedCase.isIndexCase(), MvcLink.of(controller.getContactsOfCase(caseId, null), CONTACTS))
-				.andIf(trackedCase.getStatus().equals(Status.IN_REGISTRATION) ||
-						((trackedCase.getStatus().equals(Status.REGISTERED) || trackedCase.getStatus().equals(Status.TRACKING))
-								&& trackedCase.getTrackedPerson().getAccount().isEmpty()), () -> MvcLink.of(uri, RENEW))
-				.andIf(trackedCase.getTrackedPerson().getAccount().isPresent(),MvcLink.of(controller.removeAccountOfTrackedPerson(caseId, null), ACCOUNT))
-				.andIf(trackedCase.isEligibleForTracking(), () -> MvcLink.of(uri, START_TRACKING))
+				.andIf(trackedCase.getTrackedPerson().getAccount().isEmpty(), () -> MvcLink.of(registrationController.createRegistration(caseId, null), RENEW))
+				.andIf(trackedCase.getTrackedPerson().getAccount().isPresent(),MvcLink.of(registrationController.removeAccountOfTrackedPerson(caseId, null), ACCOUNT))
+				.andIf(trackedCase.isEligibleForTracking(), () -> MvcLink.of(registrationController.createRegistration(caseId, null), START_TRACKING))
 				.andIf(trackedCase.hasQuestionnaire(),
 						() -> MvcLink.of(controller.getQuestionnaire(caseId, null), QUESTIONNAIRE))
 				.and(trackedCase.getOriginCases().stream()
